requirements_postgres.txt
pandas
holidays
scikit-learn
openpyxl
psycopg2-binary
